Move config instance building into configserver
